hook: support a --to-stdin=<path> option for testing
Expose the "path_to_stdin" API added in the preceding commit in the
"git hook run" command. We will not be using this via the command
interface in upcoming commits, but exposing this functionality makes
it easier to test the hook API.

This isn't strictly necessary as we could run
e.g. "t5407-post-rewrite-hook.sh", but building up "git hook run" to
1=1 map to the hook API makes it easier to understand and test the
API.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>