hook API: provide feed_pipe API, convert 'post-rewrite' hook to hook.h
In cases where a hook requires only a small amount of information via
stdin, it should be simple for users to provide a string_list alone. But
in more complicated cases where the stdin is too large to hold in
memory, let's instead provide a callback the users can populate line
after line.

To do this add a pipe_from_string_list() utility function to the hook
API. For now it's only used in sequencer.c, but in subsequent commits
other callers will make use of this new API.

Use this new API to convert the one remaining "post-rewrite"
invocation in sequencer.c to the hook API. This simplifies hook
invocations by not needing to put together our own 'struct
child_process'.

The signal handling that's being removed by this commit now takes
place in run-command.h:run_processes_parallel(), so it is OK to remove
it here.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>