hooks tests: don't leave "actual" nonexisting on failure
Start by creating an "actual" file in a core.hooksPath test that has
the hook echoing to the "actual" file. We later test_cmp that file to
see what hooks were run. Having that list of files be empty if we fail
instead of throwing an error about it being nonexistent makes for
friendlier output, we'll see what hooks we should have run.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>