path.c: support !repo->gitdir in strbuf_worktree_gitdir()
Change the strbuf_worktree_gitdir() API so that it can be called
outside of a repository without segfaulting.

This allows for calling it e.g. via the strbuf_git_path() ->
do_git_path() -> strbuf_worktree_gitdir() chain outside of a
repository.

We can thus use e.g. a strbuf_git_path with a "hooks/%s" format to
either get the path to the hooks directory, or an empty string in case
we can't find that path. This is preferable to the caller needing to
duplicate the logic already in strbuf_worktree_gitdir(), i.e. checking
if we have a worktree, whether we have a repository gitdir etc.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>