docs: move pack format docs to man section 5
Continue the move of existing Documentation/technical/* protocol and
file-format documentation into our main documentation space by moving
the pack-format.txt file and updating things that refer to it to link
to the new location.

I think it would be better for link-ability to split up that document
so that at least the *.rev format and multi-pack-index are discussed
in separate documents, but let's not do that now.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>