revision.h: add an "again" return value to "handle_stdin_line"
In a preceding commit we added a handle_stdin_line() callback
mechanism to the revision.c API.

That callback has the limitation that you can't see what's happened
after handle_revision_arg() has been called, i.e. what pending commits
it has added, or to act on that set of pending commits.

One way to solve this would be to make handle_revision_arg()
non-static again, and do our own line parsing, calling of it, and
"seen_dashdash" handling.

I think a better approach is to allow the callback to return a "call
me again before the next line" value. We'll then call the callback
again, and it's expected to return a "continue". The callback will be
responsible for keeping its own state of whether it's being called the
first or second time via its "stdin_line_priv".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>