bundle: allow specifying ref names in --stdin input
The bundle archive format inherently requires specifying reference
names along with the revision tips contained therein, but it doesn't
inherently require that reference names existed in that state in a
repository.

The current behavior around dwim_ref() is merely a convenience so that
a revision like "master" "master~10..master" will be listed in the
bundle as the fully-qualified. "refs/heads/master". Let's also support
manually specifying reference names along with revisions, so that
revisions like "$(git rev-parse HEAD)" or "master~10..master~5" can be
packaged in a bundle.

To do so we use the new handle_stdin_line() callback mechanism in
revision.c to intercept the --stdin input, and effectively extend
handle_revision_arg() to not only understand input like:

    <revision>

But also:

    <revision><TAB><refname>

E.g. this now works:

    printf "master~10..master~5\trefs/heads/my-recent-master-pushes\n" |
    git bundle create master-pushes.bundle --stdin

Before this change "bundle create" would have died with "Refusing to
create empty bundle" if it had gotten that far (but would earlier die
on "bad revision" if that was supplied on stdin).

This doesn't make anything possible that wasn't possible before this
change, merely a lot more convenient.

Before you could manually search/replace/edit the contents of the
bundle header if you wanted to rename/insert/edit/delete
references. In other cases such as the "master~10..master~5" case, you
would have had to create a temporary branch "tmp" whose tip was at
"master~5", and then refer to the revision range as "tmp~5..tmp".

Implementation notes:

I think the whole use of "refname_to_pending" here is rather nasty,
but could not think of a better alternative to track a
"rev_info->pending" from the outside.

The semantics are rather straightforward though: As we parse lines on
stdin we have the revision.c API parse the revision part, and it
pushes a revision (or none, or two) to "rev_info->pending".

Our "handle_stdin_line" callback then runs, sees how much (if
anything) got added to "rev_info->pending". It then extends a
string-list correspondingly such that we end up with a 1=1 positional
mapping between the "rev_info->pending" and our own
"refname_to_pending" string-list.

It results in some complexity in the stdin handler, but means that by
the time we get to write_bundle_refs() we merely need to check if
there's something in the equivalent of refname_to_pending[i] for the
revs_pending[i] we're looking at. If there is we'll use that name,
otherwise fall back on the existing dwim_ref() behavior. See the
commentary in write_bundle_after_stdin_line_again() for how we satisfy
that criteria.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>